feat: add React Router and restructure project folders
