+++ rs-react-app/package.json
@@ -17,6 +17,7 @@
     "@tailwindcss/vite": "^4.1.11",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router": "^7.7.0",
     "tailwindcss": "^4.1.11"
   },
   "devDependencies": {
