+++ rs-react-app/src/components/MainLayout.tsx
@@ -1,0 +1,11 @@
+import { Outlet } from 'react-router';
+import { Navbar } from './Navbar';
+
+export const MainLayout = () => {
+  return (
+    <main>
+      <Navbar />
+      <Outlet />
+    </main>
+  );
+};
