+++ rs-react-app/src/components/Navbar.tsx
@@ -1,0 +1,7 @@
+export const Navbar = () => {
+  return (
+    <div>
+      <h1>Navbar</h1>
+    </div>
+  );
+};
