+++ rs-react-app/src/pages/AboutPage.tsx
@@ -1,0 +1,7 @@
+export const AboutPage = () => {
+  return (
+    <div>
+      <h1>AboutPage</h1>
+    </div>
+  );
+};
