+++ rs-react-app/src/pages/ErrorPage.tsx
@@ -1,0 +1,7 @@
+export const ErrorPage = () => {
+  return (
+    <div>
+      <h1>ErrorPage</h1>
+    </div>
+  );
+};
